neumorphic update
I think that I have everything done for it now. I need to go through and clean it all up, but it looks amazing honestly! I may do something with the corner file? Idk, and then I do need to decide how I want to do the standards or if I want to do the standards for this