welcome.dart draft done
I think that I got a majority of the standard themes setup for the welcome screen. Now I need to start doing those other screens like I talked about (the signIn/signUp/forgotPassword). After I do those, then I can get started on the application itself, which we are still struggling to decide what that is supposed to be.

I think that a good way to do it though would be to just simply go with an idea all of the way. For the past few years I have done nothing but create 50% of an app before stopping and I know that if I just made a really crappy app every like two weeks, eventually I would start making some good ones. So maybe that's what I need to do here