authComponents.dart update
Getting there on the authentication page. I've been making a lot of good progress, basically just have to refactor a bunch of stuff and then actually go in and position/make it look good.

Still need to do responsiveness for both this page and welcome.dart

I also did some changes to the colors.dart file making it even smaller and easier to read