auth.dart update
Okay, I think that I finally got it all arranged how I want it. I need to go in, work on some of the mechanics, and get all of the final stuff worked out (like the theme stuff).

Then I should be able to throw something together rather quickly for the forgotPassword screen and then move onto the home screen