settings update
Making some headway on the settings UI. I think that I have the settings themselves setup properly now. I got all the required ones and then room for optional ones like options for a listview.

I need to do a bunch of cleanup on the UI and figure out what else I need to do to make it work properly and efficiently