settings update
Getting closer and closer. I kind of want to go in and finish up the forgotPassword page really quickly, that way I can focus 100% on the settings page for errors when i try to test it. I don't even remember what I have left to do on the forgotPassword page besides I guess test it and make sure all of the positions and colors are correct