forgotPassword update
Okay, I think that the forgotPassword page is getting there. We definitely still need to do some tweaking, and do a lot of work on the responsiveness idea, but I think we are at a good point. Now I kind of want to do some testing on the serttings page