auth update
Getting there on the privacy and termsAndConditions stuff. Need to do cleanup on the widget for auth (do language stuff for it). Then figure out how we want to do the opening of the files.

Then we need to do the settings page (may just do the language stuff for it at the same time)