ImageViewer updates
Getting there on the imageViewer, I need to figure out more with this scale and go from there. Honestly, if I can get this working properly, it would be great. I just want the actual image to be what the user sees, I feel like that makes sense, but most people online have just changed the image widget to display on the screen, which doesn't really make sense to me.