Management updates
Decided to go against VariableDeclaration because I started thinking, "Well where does it end?" If I have half the variables in there, do I just need to put everything in there? And everything for every method?

And then I decided to put the imageProcess inside of cameraManagement because it's taking the camera image file and flipping it first of all... Actually may make a flip method in camera and make image