Web updates
Trying to get web/iOS support done. I got web basic support done, but Im definitely going to have to test some things and then make sure that all of the userInterface stuff works (that is going to be the hardest, just figuring out how that should be handled. I feel as though once I get it done for one of them, it will make it easier for the other one)