New Idea
I was getting there on the forgotPassword, but then I thought about needing a minimumHeight and stuff. We may need to go back and do some work on the web side of stuff so that we can have the user able to do everything.

I need to figure out how we would want to format the min stuff. I would think probably same as dimensions?