Welcome Redesign part 2
We have all the themes and stuff done, but now... Now we are going to start doing something big next time.

I think that I am going to completely remove the welcome page. I want to get it to just auth, and create an animatedContainer that will size up into the auth stuff. I think that it will be very clean, will clean up some stuff in the code and will do a lot for the app.

I want to get this done tonight, then tomorrow I can do the forgotPassword cleanup.

Then after that I can do the error page design.